+++ src/plone/restapi/deserializer/dxfields.py
@@ -81,6 +81,11 @@ class DatetimeFieldDeserializer(DefaultFieldDeserializer):
         else:
             tzinfo = None
 
+        # This happens when a 'null' is posted for a non-required field.
+        if value is None:
+            self.field.validate(value)
+            return
+
         # Parse ISO 8601 string with Zope's DateTime module
         try:
             dt = DateTime(value).asdatetime()
@@ -199,7 +204,7 @@ class NamedFieldDeserializer(DefaultFieldDeserializer):
                 # We are probably pushing the contents of a previous GET
                 # That contain the read representation of the file
                 # with the 'download' key so we return the same stored file
-                return self.context.image
+                return getattr(self.field.context, self.field.__name__)
 
             content_type = value.get(u'content-type', content_type).encode(
                 'utf8')